Rank F80 M3 and 991.2 GTS
Loved the F80 M3! I wished it didn't have glasses-eyes for headlights and a giant muffler visible between its tailpipes. But driving dynamics-wise, it was so stable and grippy thanks to its long wheelbase, and everything about it felt familiar-but-better having driven a 135i for years.

The 991.2 GTS felt like a race car. But its steering was too precise (it felt sterile), and it was stressful not keeping grip on not-perfectly-maintained backroads.